This introduces some new functions to handle the specpdl.  The basic idea is that when a thread loses the interpreter lock, it will unbind the bindings it has put in place.  Then when a thread acquires the lock, it will restore its bindings.
This code reuses an existing empty slot in struct specbinding to store
the current value when the thread is "swapped out".

This approach performs worse than my previously planned approach.
However, it was one I could implement with minimal time and
brainpower.  I hope that perhaps someone else could improve the code
once it is in.