Don't explicitly bind mouse-1.
As Michael Heerdegen <michael_heerdegen@web.de> says:

However, also binding mouse-1 is not necessary.  Depending on the
setting of `mouse-1-click-follows-link', which is non-nil by default,
mouse-1 already follows links - see its doc.  We should leave it up to
the user if he wants to use mouse-1 for following or not.